Release/2020.3.0 nginx

+++ vm-docker-compose/config.sh
@@ -1,9 +1,16 @@
 #!/usr/bin/env bash
 
 { # make sure that the entire script is downloaded #
+    VERSION="1.0.0"
+
+    DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+
+    # Shared code
+    source ${DIR}/shared.sh
+
     check_installed_programs() {
 
-        for i in envsubst grep tr cut ; do
+        for i in docker-compose envsubst sed grep tr cut ; do
             if ! [ -x "$(command -v ${i})" ]; then
                 error "Error: ${i} is not installed." >&2
                 exit 1
@@ -35,7 +42,147 @@
             esac
         done
 
-        echo "Protocol : " ${PROTOCOL}
+        info "Protocol : ${PROTOCOL}" 
+
+    }
+
+
+    input_SYSMON() {
+
+        PS3='Do you want to install the system monitoring tools : '
+        echo
+
+        local _options=("y" "n" "exit")
+        select SELECT in "${_options[@]}"
+        do
+            case $SELECT in
+                "y")
+                    export SYSMON=y
+                    break
+                    ;;
+                "n")
+                    export SYSMON=n
+                    break
+                    ;;
+                "exit")
+                    exit 1
+                    ;;
+                *) error "invalid option $REPLY";;
+            esac
+        done
+        
+        info "System Monitoring Tools : ${SYSMON}" 
+
+    }
+    input_SUPPORT() {
+
+        PS3='Do you want to install the remote support tools : '
+        echo
+
+        local _options=("y" "n" "exit")
+        select SELECT in "${_options[@]}"
+        do
+            case $SELECT in
+                "y")
+                    export SUPPORT=y
+                    break
+                    ;;
+                "n")
+                    export SUPPORT=n
+                    break
+                    ;;
+                "exit")
+                    exit 1
+                    ;;
+                *) error "invalid option $REPLY";;
+            esac
+        done
+        
+        info "Support Tools : ${SUPPORT}" 
+
+    }
+
+    input_INGRESS_PROTOCOL() {
+
+        PS3='Do you want to use INGRESS with SSL '
+        echo
+
+        local _options=("y" "n" "exit")
+        select SELECT in "${_options[@]}"
+        do
+            case $SELECT in
+                "y")
+                    export INGRESS_SSL=y
+                    export PROTOCOL_INGRESS=https
+                    input_CERTBOT_DOCKER
+                    break
+                    ;;
+                "n")
+                    export INGRESS_SSL=n
+                    export PROTOCOL_INGRESS=http
+                    break
+                    ;;
+                "exit")
+                    exit 1
+                    ;;
+                *) error "invalid option $REPLY";;
+            esac
+        done
+
+        info "INGRESS with SSL certificate : ${INGRESS_SSL}" 
+        info "INGRESS Protocol : ${PROTOCOL_INGRESS}" 
+
+    }
+
+    input_NGINX() {
+
+        PS3='Do you want to run NGINX as a docker container : '
+        echo
+
+        local _options=("y" "n" "exit")
+        select SELECT in "${_options[@]}"
+        do
+            case $SELECT in
+                "y")
+                    export NGINX_DOCKER=y
+                    break
+                    ;;
+                "n")
+                    export NGINX_DOCKER=n
+                    break
+                    ;;
+                "exit")
+                    exit 1
+                    ;;
+                *) error "invalid option $REPLY";;
+            esac
+        done
+
+    }
+
+    input_CERTBOT_DOCKER() {
+
+        PS3='Do you want to run CERTBOT to maintain your LETSENCYPT certificate : '
+        echo
+
+        local _options=("y" "n" "exit")
+        select SELECT in "${_options[@]}"
+        do
+            case $SELECT in
+                "y")
+                    export CERTBOT_DOCKER=y
+                    break
+                    ;;
+                "n")
+                    export CERTBOT_DOCKER=n
+                    break
+                    ;;
+                "exit")
+                    exit 1
+                    ;;
+                *) error "invalid option $REPLY";;
+            esac
+        done
 
     }
 
@@ -70,24 +217,23 @@
         ## ip, dns, ingress
         if [[ ${DEPLOYMENT} == "ip" ]] ; then
             input_EXTERNAL_IP
-            echo "External IP : " $EXTERNAL_IP
+            info "External IP : $EXTERNAL_IP" 
         fi
 
         ## ip, dns, ingress
         if [[ ${DEPLOYMENT} == "dns" ]] ; then
             input_DNS
-            echo "DNS : " $DNS
+            
         fi
                 ## ip, dns, ingress
         if [[ ${DEPLOYMENT} == "ingress" ]] ; then
-            input_INTERNAL_IP
+            # input_INTERNAL_IP
             input_DNS
 
-            echo "DNS : " ${DEPLOY_URL}
-            echo "Internal IP : " ${INTERNAL_IP}
+            info "DNS : ${DEPLOY_URL}" 
         fi
 
-        echo "Deployment : " ${DEPLOYMENT}
+        info "Deployment :${DEPLOYMENT} "
 
     }
 
@@ -121,9 +267,9 @@
                     EXTERNAL_IP=$var1
                     break
                 else
-                    echo
-                    echo ' Invalid IP address ' $var1 
-                    echo
+                    error
+                    error ' Invalid IP address ' $var1 
+                    error
                 fi
             fi
         done
@@ -132,7 +278,7 @@
     input_INTERNAL_IP() {
 
         while true; do
-            echo
+            info
             read -p "Enter internal ip address: or EXIT " var1
 
             if [[ $var1 == "exit" ]]
@@ -144,16 +290,16 @@
                     INTERNAL_IP=$var1
                     break
                 else
-                    echo
-                    echo ' Invalid IP address ' $var1 
-                    echo
+                    error
+                    error ' Invalid IP address ' $var1 
+                    error
                 fi
             fi
         done
-        echo
+        info
+        info "Internal IP : ${INTERNAL_IP}" 
 
     }
-
 
     input_DNS() {
         while true; do
@@ -174,35 +320,8 @@
                 # fi
             fi
         done
-    }
-
-    input_ADVANCED_OPTIONS() {
-        PS3='Select the Type of Deployment for your deployment : '
-        echo
-
-        local _options=("ip" "dns" "ingress" "exit")
-        select SELECT in "${_options[@]}"
-        do
-            case $SELECT in
-                "ip")
-                    export DEPLOYMENT=ip
-                    break
-                    ;;
-                "dns")
-                    export DEPLOYMENT=dns
-                    break
-                    ;;
-                "ingress")
-                    export DEPLOYMENT=ingress
-                    break
-                    ;;
-                "exit")
-                    exit 1
-                    ;;
-                *) error "invalid option $REPLY";;
-            esac
-        done
-
+        info
+        info "DEPLOY_URL : $DEPLOY_URL" 
     }
 
     main() {
@@ -211,30 +330,83 @@
 
         input_DEPLOYMENT
 
-        if [[ "${DEPLOYMENT}" == "ip" || "${DEPLOYMENT}" == "dns" ]]; then
-            input_PROTOCOL
-        else
-            if [[ "${DEPLOYMENT}" == "ingress" ]]; then
-            # DEPLOYMENT 
-             PROTOCOL="http"
-            fi
+        if [[ "${DEPLOYMENT}" == "ip" ]]; then
+            PROTOCOL="http"
         fi
-    
-        # input_ADVANCED_OPTIONS
+
+        if [[ "${DEPLOYMENT}" == "ingress" ]]; then
+            PROTOCOL="http"
+            input_INGRESS_PROTOCOL
+        fi
+
+        if [[ "${DEPLOYMENT}" == "dns" ]]; then
+            # Can run with or w/o SSL
+            input_PROTOCOL
+        fi
+        
+        input_SUPPORT
+        input_SYSMON
 
         export DEPLOYMENT=${DEPLOYMENT}
         export PROTOCOL=${PROTOCOL}
         export DEPLOY_URL=${DEPLOY_URL}
 
-        echo "Backup existing .env file"
-        cp .env .env.before_config
-        envsubst < .env.template > .env
+        FILE=.env
+        if [[ -f "$FILE" ]]; then
+            info "Backup existing .env file"
+            cp .env .env.before_config
+        fi
+
+        FILE=docker-compose.yml
+        if [[ -f "$FILE" ]]; then
+            info "Backup existing docker-compose file"
+            cp docker-compose.yml docker-compose.yml.before_config
+        fi
+
+        info "Creating docker-compose environment"
+        envsubst < ./templates/.env.template > .env
+        cp ./templates/.docker-compose.template ./docker-compose.yml
 
         if [[ "${DEPLOYMENT}" == "ingress" ]]; then
-            echo "Configure NGINX"
-            # DEPLOYMENT 
-             . "./nginx_create_site.sh" ${INTERNAL_IP} ${DEPLOY_URL}
+            input_NGINX
+
+            if [[ "${NGINX_DOCKER}" == "y" ]]; then
+                info "Configure NGINX in docker "
+
+                if [[ "${PROTOCOL_INGRESS}" == "http" ]]; then
+                    readonly REPL_DEPLOY_URL=domain
+                    cp ./templates/.nginx-docker.template ./nginx/sites/${DEPLOY_URL}
+                    sed -i "s/${REPL_DEPLOY_URL}/${DEPLOY_URL}/g" ./nginx/sites/${DEPLOY_URL}
+                else 
+                    readonly REPL_DEPLOY_URL=domain
+                    cp ./templates/.nginx-docker-ssl.template ./nginx/sites/${DEPLOY_URL}
+                    sed -i "s/${REPL_DEPLOY_URL}/${DEPLOY_URL}/g" ./nginx/sites/${DEPLOY_URL}
+                fi
+
+                cat ./templates/.docker-compose-nginx.template >> ./docker-compose.yml
+                
+                if [[ "${CERTBOT_DOCKER}" == "y" ]]; then
+                    cat ./templates/.docker-compose-certbot.template >> ./docker-compose.yml
+                fi
+
+            else
+                info "Configure NGINX installed outside of the docker network"
+                # DEPLOYMENT 
+                input_INTERNAL_IP
+                . "./nginx_create_site.sh" ${INTERNAL_IP} ${DEPLOY_URL}
+            fi
         fi
+
+        if [[ "${SUPPORT}" == "y" ]]; then
+            info "Configure Support tools"
+            cat ./templates/.docker-compose-portainer.template >> ./docker-compose.yml
+        fi
+
+        if [[ "${SYSMON}" == "y" ]]; then
+            info "Configure System monitoring"
+            cat ./templates/.docker-compose-sysmon.template >> ./docker-compose.yml
+        fi
+
     }
 
   main @1
